Update color scheme

+++ process_bags.py
@@ -2,12 +2,19 @@ import os
 import pickle
 import src.bag_utils as utils
 
-BAGS_FOLDER = '/data/deployment_nov2023'
-TMP_FILE = os.path.join(BAGS_FOLDER, 'processed_data.pkl')
-
 
 if __name__ == '__main__':
-    gt_path = os.path.join(BAGS_FOLDER, 'GT_Nov2023.csv')
-    data = utils.process_bags(bags_folder=BAGS_FOLDER, gt_path=gt_path)
-    with open(TMP_FILE, 'wb') as f:
+    bags_folder = os.environ.get('BAGS_PATH')
+    if not bags_folder:
+        raise ValueError('BAGS_PATH is not set')
+    if not os.path.isdir(bags_folder):
+        raise ValueError(f'Unable to locate BAGS_PATH: {bags_folder}')
+    
+    gt_path = os.environ.get('GT_PATH')
+    if gt_path and not os.path.isfile(gt_path):
+        raise ValueError(f'Unable to locate GT_PATH: {gt_path}')
+
+    tmp_file = os.path.join(bags_folder, 'processed_data.pkl')
+    data = utils.process_bags(bags_folder=bags_folder, gt_path=gt_path)
+    with open(tmp_file, 'wb') as f:
         pickle.dump(data, f)
